Removes `framework.WithImportByGlobalARN`
